+++ Cargo.toml
@@ -29,7 +29,7 @@ quote = "1.0.37"
 rand = { version = "0.8.5", default-features = false }
 rand_chacha = { version = "0.3.1", default-features = false }
 sha2 = { version = "0.10.9", default-features = false }
-rayon = "1.10.0"
+rayon = "1.11.0"
 serde = { version = "1.0.216", features = ["derive"] }
 serde_json = "1.0.137"
 stark-rings = { git = "https://github.com/NethermindEth/stark-rings.git", branch = "main", default-features = false }
