+++ crates/latticefold-plus/src/sp1_oneproof_api.rs
@@ -1,0 +1,292 @@
+//! API wrapper for the `examples/lf_plus_sp1_oneproof.rs` harness.
+//!
+//! This exists so downstream crates (like PVUGC) can call the SP1→LF+→WE-gate path **in-process**
+//! without spawning binaries or scraping logs.
+//!
+//! This is research plumbing and intentionally returns only a small structured summary.
+
+#![cfg(feature = "we_gate")]
+
+use cyclotomic_rings::rings::{FrogPoseidonConfig as PC, GetPoseidonParams};
+use cyclotomic_rings::rings::FrogRing64 as R;
+use ark_ff::{BigInteger, PrimeField};
+use latticefold::commitment::AjtaiCommitmentScheme;
+use stark_rings::PolyRing;
+use stark_rings_linalg::SparseMatrix;
+
+use rand::{rngs::StdRng, RngCore, SeedableRng};
+use sha2::{Digest, Sha256};
+use std::sync::Arc;
+
+use crate::we_statement::{we_statement_hash_lf_plus, LFP_WE_GATE_DIGEST_V1};
+
+/// Structured output for downstream wiring (no stdout parsing).
+#[derive(Clone, Debug)]
+pub struct Sp1OneProofWeGateOutput {
+    pub stmt_digest: [u8; 32],
+    pub lock_coin_seed: [u8; 32],
+    /// Best-effort dR1CS **assignment tail** for downstream lock scaffolding.
+    ///
+    /// Concretely: `assignment[public_len..]` (i.e., the non-public variables) converted to `u64`
+    /// via canonical reduction in the small field.
+    ///
+    /// This is **not** the final Architecture bounded-integer witness representation.
+    pub dr1cs_assignment_tail_u64: Vec<u64>,
+}
+
+fn hex32(bytes: &[u8; 32]) -> String {
+    const HEX: &[u8; 16] = b"0123456789abcdef";
+    let mut out = String::with_capacity(64);
+    for &b in bytes {
+        out.push(HEX[(b >> 4) as usize] as char);
+        out.push(HEX[(b & 0x0f) as usize] as char);
+    }
+    out
+}
+
+/// Run the SP1 oneproof + WE gate path from on-disk artifacts.
+///
+/// Inputs are the same formats the example consumes:
+/// - `r1lf_path`: SP1 shrink verifier `.r1lf`
+/// - `witness_path`: SP1 witness bundle `.bundle` ("SP1W")
+pub fn run_sp1_oneproof_we_gate_from_files(
+    r1lf_path: &str,
+    witness_path: &str,
+) -> Result<Sp1OneProofWeGateOutput, String> {
+    // Keep defaults aligned with the example (but avoid printing / panics).
+    let chunk_size: usize = std::env::var("CHUNK_SIZE")
+        .ok()
+        .and_then(|s| s.parse().ok())
+        .unwrap_or(1 << 20);
+    let pad_cols_to_multiple_of: usize = std::env::var("PAD_COLS")
+        .ok()
+        .and_then(|s| s.parse().ok())
+        .unwrap_or(256);
+
+    let cache = crate::sp1_r1lf::open_sp1_r1lf_chunk_cache::<R>(
+        r1lf_path,
+        chunk_size,
+        pad_cols_to_multiple_of,
+    )
+    .map_err(|e| format!("open_sp1_r1lf_chunk_cache: {e}"))?;
+
+    // Materialize full (A,B,C) as SparseMatrix<F> (constant-coeff) by concatenating chunk rows.
+    type F = <R as PolyRing>::BaseRing;
+    let total_rows = cache.num_chunks * chunk_size;
+    let mut a_rows: Vec<Vec<(F, usize)>> = Vec::with_capacity(total_rows);
+    let mut b_rows: Vec<Vec<(F, usize)>> = Vec::with_capacity(total_rows);
+    let mut c_rows: Vec<Vec<(F, usize)>> = Vec::with_capacity(total_rows);
+    for chunk_idx in 0..cache.num_chunks {
+        let [a, b, c] = cache.read_chunk(chunk_idx).map_err(|e| format!("read_chunk: {e}"))?;
+        let conv = |m: stark_rings_linalg::SparseMatrix<F>| m.coeffs;
+        let (ar, br, cr) = (conv(a), conv(b), conv(c));
+        a_rows.extend(ar);
+        b_rows.extend(br);
+        c_rows.extend(cr);
+    }
+    let m_a = SparseMatrix::<F> { nrows: total_rows, ncols: cache.ncols, coeffs: a_rows };
+    let m_b = SparseMatrix::<F> { nrows: total_rows, ncols: cache.ncols, coeffs: b_rows };
+    let m_c = SparseMatrix::<F> { nrows: total_rows, ncols: cache.ncols, coeffs: c_rows };
+
+    let bundle = crate::sp1_witness_io::load_sp1_witness_any(witness_path, cache.stats.num_vars)
+        .map_err(|e| format!("load_sp1_witness_any: {e}"))?;
+    let (w_u64, _base_len, _aux_len) = (bundle.witness, bundle.base_len, bundle.aux_len);
+
+    // Map u64 witness -> Frog base field scalars once (centered embedding mod p_bb).
+    let p_bb = cache.stats.p_bb;
+    let mut w_host: Vec<F> = Vec::with_capacity(w_u64.len());
+    for &x in &w_u64 {
+        w_host.push(babybear_u64_to_centered_host::<F>(x, p_bb));
+    }
+
+    // Public inputs from the SP1 export are witness[1..=l_pub].
+    let l_pub = cache.stats.num_public;
+    if l_pub == 0 {
+        return Err("SP1 R1LF exports num_public=0; statement binding requires public inputs".to_string());
+    }
+    if w_host.len() < 1 + l_pub {
+        return Err(format!(
+            "witness too short for declared public inputs: w_len={} need_at_least={}",
+            w_host.len(),
+            1 + l_pub
+        ));
+    }
+    type BFSmall = <<R as PolyRing>::BaseRing as ark_ff::Field>::BasePrimeField;
+    let mut public_inputs: Vec<BFSmall> = w_host[1..1 + l_pub].to_vec();
+    // Debug knob parity with the historical harness: mutate a public input and ensure things fail.
+    if !public_inputs.is_empty()
+        && std::env::var("FLIP_PUBLIC_INPUT0").ok().as_deref() == Some("1")
+    {
+        public_inputs[0] = <BFSmall as ark_ff::Field>::ONE - public_inputs[0];
+    }
+
+    // Mirror the example: deterministic Ajtai commitment scheme (system parameter) with prefix exposure.
+    let kappa_expose: usize = 8;
+    let kappa_random: usize = 8;
+    let kappa: usize = kappa_expose + kappa_random;
+    const AJTAI_SEED: [u8; 32] = *b"LFP_SP1_AJTAI_SEED_V1_0000000000";
+    let ajtai = AjtaiCommitmentScheme::<R>::seeded_with_exposed_prefix(
+        b"lf_plus_ajtai",
+        AJTAI_SEED,
+        kappa,
+        cache.ncols,
+        kappa_expose,
+        1, // expose witness columns 1..=8 (skip shared ONE at column 0)
+    );
+
+    // Build ComR1CS from f0 (base witness) and keep matrices Arc for the prover.
+    let mut f0 = w_host.clone();
+    f0.truncate(cache.stats.num_vars);
+    let f0: Arc<Vec<F>> = Arc::new(f0);
+    let r1cs = latticefold::arith::r1cs::R1CS::<F> { l: l_pub, A: m_a, B: m_b, C: m_c };
+    let cr1cs =
+        crate::r1cs::ComR1CSBase::<R>::from_f0_seeded_base(r1cs, f0, l_pub, &ajtai);
+    let m0 = cr1cs.x.matrices_arc_base();
+
+    // WE params derived from SP1 cache + chosen kappa.
+    let we_params = crate::sp1_r1lf::sp1_default_we_params_for_r1lf_cache::<R>(
+        &cache,
+        kappa as u64,
+        m0.len() as u64,
+    )
+    .map_err(|e| format!("sp1_default_we_params_for_r1lf_cache: {e}"))?;
+    let dparams = crate::rgchk::DecompParameters {
+        b: (we_params.decomp_b as u128),
+        k: (we_params.k as usize),
+        l: (we_params.l as usize),
+    };
+    let lin_params = crate::lin::LinParameters { kappa, decomp: dparams };
+
+    // Choose decomposition base B > sqrt(q), rounded up to a power of 2 (same as example).
+    fn isqrt_u128(x: u128) -> u128 {
+        let mut lo: u128 = 0;
+        let mut hi: u128 = 1u128 << 64;
+        while lo + 1 < hi {
+            let mid = (lo + hi) >> 1;
+            if mid.saturating_mul(mid) <= x {
+                lo = mid;
+            } else {
+                hi = mid;
+            }
+        }
+        lo
+    }
+    fn next_pow2_u128(x: u128) -> u128 {
+        if x <= 1 {
+            return 1;
+        }
+        let p = 128 - (x - 1).leading_zeros() as u32;
+        1u128 << p
+    }
+    let q_u128: u128 = <F as PrimeField>::MODULUS.0[0] as u128;
+    let b_min = isqrt_u128(q_u128) + 1;
+    let b_decomp: u128 = next_pow2_u128(b_min);
+    let pparams = crate::plus::PlusParameters { lin: lin_params, B: b_decomp };
+
+    // Statement digest binds vk, r1cs digest, gate-id, params, and public inputs.
+    if bundle.r1lf_digest != cache.stats.digest {
+        return Err(format!(
+            "witness bundle r1lf_digest does not match SP1_R1LF cache: bundle=0x{} cache=0x{}",
+            hex32(&bundle.r1lf_digest),
+            hex32(&cache.stats.digest)
+        ));
+    }
+    let (vk_hash, _committed_values_digest) = bundle.public_inputs;
+    let r1cs_digest = cache.stats.digest;
+    let stmt_digest =
+        we_statement_hash_lf_plus::<R>(vk_hash, r1cs_digest, LFP_WE_GATE_DIGEST_V1, &we_params, &public_inputs);
+
+    // Derive one representative coin seed (j=0) like the example.
+    const ARMER_SEED: [u8; 32] = *b"LFP_ARMER_SEED_V1_00000000000000";
+    let lock_j: u64 = 0;
+    let lock_coin_seed: [u8; 32] = {
+        let mut h = Sha256::new();
+        h.update(b"LFP_LOCK_COIN_V1");
+        h.update(&ARMER_SEED);
+        h.update(&stmt_digest);
+        h.update(&lock_j.to_le_bytes());
+        h.finalize().into()
+    };
+
+    // Plus prover.
+    let mut prover = crate::plus::PlusProverSparseBase::init_seeded_base(
+        ajtai.clone(),
+        m0.clone(),
+        1,
+        pparams.clone(),
+        crate::transcript::PoseidonTranscript::empty::<PC>(),
+    );
+    let proof = prover.prove_sparse_base(std::slice::from_ref(&cr1cs), &public_inputs);
+
+    // Record verifier trace and build WE gate dR1CS; then extract witness tail.
+    let poseidon_cfg = PC::get_poseidon_config();
+    let mut rec = crate::recording_transcript::TracePoseidonTranscript::<R>::empty::<PC>();
+    for b in &public_inputs {
+        rec.absorb_field_element(b);
+    }
+    for lp in &proof.lproof {
+        lp.verify(&mut rec);
+    }
+    // Skip native prefix binding check only if explicitly requested.
+    let skip_prefix_check =
+        std::env::var("LFP_SKIP_PREFIX_BINDING_CHECK").ok().as_deref() == Some("1");
+    let bind_prefix: &[BFSmall] = if skip_prefix_check {
+        &[]
+    } else {
+        public_inputs
+            .get(0..8)
+            .ok_or_else(|| "expected at least 8 public inputs for Ajtai prefix exposure".to_string())?
+    };
+    proof
+        .cmproof
+        .verify_with_mlen(m0.len(), &mut rec, bind_prefix)
+        .map_err(|e| format!("cm proof verify: {e:?}"))?;
+    let trace = rec.trace().clone();
+
+    let out = crate::we_gate_arith::build_we_dr1cs_for_plus_proof::<R>(
+        &poseidon_cfg,
+        &trace,
+        &we_params,
+        &public_inputs,
+        &proof,
+        m0.len(),
+        b_decomp,
+    )
+    .map_err(|e| format!("build_we_dr1cs_for_plus_proof: {e}"))?;
+
+    // Extract the WE gate witness tail (excluding public) and convert to u64 (canonical).
+    let public_len = out.public_len;
+    let tail = &out.assignment[public_len..];
+    let mut dr1cs_assignment_tail_u64: Vec<u64> = Vec::with_capacity(tail.len());
+    for x in tail {
+        let bi = x.into_bigint();
+        let le = bi.to_bytes_le();
+        let mut buf = [0u8; 8];
+        for (i, b) in le.iter().take(8).enumerate() {
+            buf[i] = *b;
+        }
+        dr1cs_assignment_tail_u64.push(u64::from_le_bytes(buf));
+    }
+
+    // Debug-only: keep deterministic footprints if you need to compare runs.
+    let _ = hex32(&stmt_digest);
+
+    Ok(Sp1OneProofWeGateOutput {
+        stmt_digest,
+        lock_coin_seed,
+        dr1cs_assignment_tail_u64,
+    })
+}
+
+#[inline]
+fn babybear_u64_to_centered_host<F: ark_ff::Field>(x: u64, p_bb: u64) -> F {
+    debug_assert!(p_bb > 1);
+    let half = p_bb / 2;
+    if x > half {
+        let neg = p_bb - x;
+        -F::from(neg)
+    } else {
+        F::from(x)
+    }
+}
+
